Move `hash_engine_lib` to `experiment-control`
