+++ packages/engine/bin/cli/src/main.rs
@@ -10,8 +10,8 @@ use std::{
 
 use clap::{AppSettings, Parser};
 use error_stack::{IntoReport, Result, ResultExt};
+use experiment_control::environment::init_logger;
 use experiment_structure::{ExperimentType, Manifest};
-use hash_engine_lib::experiment::environment::init_logger;
 use orchestrator::{Experiment, ExperimentConfig, Server};
 
 /// Arguments passed to the CLI
