+++ packages/engine/lib/experiment-control/src/controller/config.rs
@@ -1,7 +1,7 @@
 use execution::package::simulation::output::persistence::local::LocalPersistenceConfig;
 use serde::{Deserialize, Serialize};
 
-use crate::experiment::{environment::Environment, Error, Result};
+use crate::{environment::Environment, Error, Result};
 
 pub const OUTPUT_PERSISTENCE_KEY: &str = "output_persistence";
