+++ packages/engine/lib/experiment-control/src/environment/args.rs
@@ -3,54 +3,63 @@ use std::path::PathBuf;
 use clap::{AppSettings, Parser};
 use execution::package::experiment::ExperimentId;
 
-use crate::experiment::environment::{LogFormat, LogLevel, OutputLocation};
+use crate::environment::{LogFormat, LogLevel, OutputLocation};
 
 /// Arguments passed to hEngine
 #[derive(Debug, Parser)]
-#[clap(about, version, author)]
-#[clap(global_setting(AppSettings::PropagateVersion))]
-#[clap(setting(AppSettings::UseLongFormatForHelpSubcommand))]
+#[cfg_attr(feature = "clap", clap(about, version, author))]
+#[cfg_attr(feature = "clap", clap(global_setting(AppSettings::PropagateVersion)))]
+#[cfg_attr(
+    feature = "clap",
+    clap(setting(AppSettings::UseLongFormatForHelpSubcommand))
+)]
 /// Run the engine.
 pub struct Args {
     /// The unique identifier of the experiment, as a valid v4 UUID
-    #[clap(short, long, default_value_t = ExperimentId::generate())]
+    #[cfg_attr(feature = "clap", clap(short, long, default_value_t = ExperimentId::generate()))]
     pub experiment_id: ExperimentId,
 
     /// nng URL that the orchestrator is listening on
-    #[clap(short, long, default_value = "")]
+    #[cfg_attr(feature = "clap", clap(short, long, default_value = ""))]
     pub orchestrator_url: String,
 
     /// nng URL to listen on
-    #[clap(short, long, default_value = "")]
+    #[cfg_attr(feature = "clap", clap(short, long, default_value = ""))]
     pub listen_url: String,
 
     /// Number of workers to run in parallel.
     ///
     /// Defaults to the number of logical CPUs available in order to maximize performance.
-    #[clap(short = 'w', long, default_value_t = num_cpus::get(), env = "HASH_WORKERS")]
+    #[cfg_attr(feature = "clap", clap(short = 'w', long, default_value_t = num_cpus::get(), env = "HASH_WORKERS"))]
     pub num_workers: usize,
 
     /// Output format emitted to the output location.
-    #[clap(long, default_value = "pretty", arg_enum, env = "HASH_LOG_FORMAT")]
+    #[cfg_attr(
+        feature = "clap",
+        clap(long, default_value = "pretty", arg_enum, env = "HASH_LOG_FORMAT")
+    )]
     pub log_format: LogFormat,
 
     /// Logging verbosity to use. If not set `RUST_LOG` will be used
-    #[clap(long, arg_enum)]
+    #[cfg_attr(feature = "clap", clap(long, arg_enum))]
     pub log_level: Option<LogLevel>,
 
     /// Output location where to emit logs.
     ///
     /// Can be `stdout`, `stderr` or any file name. Relative to `--log-folder` if a file is
     /// specified.
-    #[clap(long, default_value = "stderr")]
+    #[cfg_attr(feature = "clap", clap(long, default_value = "stderr"))]
     pub output: OutputLocation,
 
     /// Logging output folder.
-    #[clap(long, default_value = "./log")]
+    #[cfg_attr(feature = "clap", clap(long, default_value = "./log"))]
     pub log_folder: PathBuf,
 
     /// The size at which the engine aims to split a group of agents
-    #[clap(long, default_value = "100000", env = "ENGINE_TARGET_MAX_GROUP_SIZE")]
+    #[cfg_attr(
+        feature = "clap",
+        clap(long, default_value = "100000", env = "ENGINE_TARGET_MAX_GROUP_SIZE")
+    )]
     pub target_max_group_size: usize,
 
     /// Heap size in megabytes of the V8 runtime in each JavaScript runner under which garbage
@@ -65,7 +74,7 @@ pub struct Args {
     /// garbage collector.
     ///
     /// Defaults to V8's `initial_heap_size` default.
-    #[clap(long)]
+    #[cfg_attr(feature = "clap", clap(long))]
     pub js_runner_initial_heap_constraint: Option<usize>,
 
     /// Maximum size in megabytes of the V8 heap in each JavaScript runner. See "--num-workers" to
@@ -76,7 +85,7 @@ pub struct Args {
     /// crashes.
     ///
     /// Defaults to V8's `max_heap_size` default.
-    #[clap(long)]
+    #[cfg_attr(feature = "clap", clap(long))]
     pub js_runner_max_heap_size: Option<usize>,
 }
