+++ packages/engine/lib/experiment-structure/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.0.0"
 edition = "2021"
 
 [dependencies]
-stateful = { path = "../stateful" }
-execution = { path = "../execution" }
+stateful = { path = "../stateful", default-features = false }
+execution = { path = "../execution", default-features = false }
 
 error-stack = "0.1.1"
