+++ packages/engine/lib/memory/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
 name = "memory"
-version = "0.1.0"
+version = "0.0.0"
 edition = "2021"
 
 [dependencies]
-flatbuffers_gen = { path = "../flatbuffers_gen" }
+flatbuffers_gen = { path = "../flatbuffers_gen", default-features = false }
 
 arrow = { version = "10.0.0", default-features = false, features = ["ipc"] }
 flatbuffers = "2.1.1"
