--- packages/engine/src/experiment/controller/mod.rs
@@ -1,8 +1,0 @@
-pub mod config;
-pub mod controller;
-pub mod run;
-pub mod sim_configurer;
-
-// TODO: Currently unused, left in in case Rust runner will use it. Can safely be removed
-//   afterwards.
-// pub mod utils;
